% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rna.foldchange.R
\name{rna.foldchange}
\alias{rna.foldchange}
\title{Compute log2-fold changes}
\usage{
rna.foldchange(expr, ref, samp)
}
\arguments{
\item{expr}{numeric matrix or data.frame consisting of gene expression values}

\item{ref}{numeric vector. Index of reference samples in the array}

\item{samp}{numeric vector. Index of case samples in the array}
}
\description{
Fold changes are computed between sample and reference arrays
}

