% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rna.pathwayttest.R
\name{rna.pathwayttest}
\alias{rna.pathwayttest}
\title{Pathway t-test}
\usage{
rna.pathwayttest(expr, gs, genelist, ref, samp, set.size = c(10, 500))
}
\arguments{
\item{expr}{numeric matrix or data.frame consisting of gene expression values}

\item{gs}{list. gene set collection}

\item{genelist}{character vector. Names of genes in the dataset.
Lenght of genelist and n rows of expr must match}

\item{ref}{numeric vector. Index of reference samples in the array}

\item{samp}{numeric vector. Index of case samples in the array}

\item{set.size}{vector. First and second elements specify the minimum and
maximum numbers of gene per pathway respectively}
}
\description{
Perform two-sample t-test to evaluate differential expression of pathways
obtained from a gene set
}

